feat: LadderTest -> 테스트 분리
